requirements.txt
python-dotenv
streamlit
langchain-google-genai
